fix(pumbility): stop truncating per chart
The reported bug. A session card read 17,195 while the PUMBILITY page
read 17,173 for the same pool on the same account, and the 22 between
them was not staleness or the formula -- the two rounded at different
points. PlayerRatingSaga summed fifty doubles and truncated once;
PumbilityPageSaga truncated each of the fifty and then summed, so the
page lost the sum of fifty discarded fractions and could only ever read
low. Both now carry the value whole, and so does everything that was
doing the same thing elsewhere:

- Phoenix2TitleList gated its ladders on a truncated pool, so a rung
  inside the discarded fraction read as unreached on the PUMBILITY page
  while the title system had already awarded it.
- PumbilityProjectionSaga truncated every projected gain.
- LeaderboardHubSaga truncated each board row before summing the top 50,
  and CoOpBoardCalculator.Rating returned an int.
- PumbilityAttribution rounded each per-chart gain to whole and dropped
  anything under a point, which is also why a sub-point gain could never
  be shown -- it did not survive to reach a badge.
- The carryover panel rounded its own totals to 2dp in the handler.

TiedAtBar compared pool values to the bar with ==. On doubles that is a
coin toss, so the "traffic jam at the bar" count would have quietly gone
to zero; it now compares at the precision the page prints.

One public change: api/v2 OfficialPlacementDto.ComputedRating is a
double rather than an int. It is a PUMBILITY value and the same payload
has always carried Pumbility as a decimal. No golden covers it.

Nothing is reformatted -- every site still renders N0, so the page and
the card now agree at 17,195. Suites green: 3328 / 736 / 148.

+++ ScoreTracker/ScoreTracker.OfficialMirror/Contracts/OfficialLeaderboardRecords.cs
@@ -91,7 +91,7 @@ public sealed record OfficialPlayerHistoryPoint(DateTimeOffset At, decimal? Pumb
 /// <summary>One chart on a player's list — mirrored board rows only, every one with a Place.</summary>
 [ExcludeFromCodeCoverage]
 public sealed record OfficialPlayerChartRecord(Guid ChartId, int? Place, int? PlaceDelta, int Score,
-    int ComputedRating, bool IsSupplemented = false);
+    double ComputedRating, bool IsSupplemented = false);
 
 /// <summary>What the supplemented reading added to the latest sealed snapshot.</summary>
 [ExcludeFromCodeCoverage]
